add arduino test sketch for sound and mic
This is the very limited test code I used to get the i2s mics up and running. There’s not much here, as I was changing things on the fly and not being terribly careful about maintaining history.

This branch should be deleted once this code is purposed. This sketch should not be in any version of the final product.